+++ CD/CD/Models/Mark.cs
@@ -7,8 +7,8 @@ namespace CD.Models
         public Guid SubjectID { get; set; }
         public Guid MarkID { get; set; }
         public string MarkName { get; set; }
-        public decimal Result { get; set; }
+        public double Result { get; set; }
         public string Category { get; set; }
-        public int Weight { get; set; }
+        public double Weight { get; set; }
     }
 }
